+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## master
+### Added
+- Add `error::last_os_error()` overloads to create an error from `errno`.
 
 ## 0.4.0 - 2020-09-01
 ### Added